requirements.txt
alembic==1.4.0
aniso8601==8.0.0
astroid==2.3.3
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cmake==3.16.3
dlib==19.19.0
dnspython==1.16.0
eventlet==0.25.1
expiringdict==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==3.3.0
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
greenlet==0.4.15
idna==2.9
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
lazy-object-proxy==1.4.3
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.22.2
mccabe==0.6.1
monotonic==1.5
numpy==1.18.1
opencv-python-headless==4.2.0.32
passlib==1.7.2
Pillow==7.0.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
typing==3.7.4.1
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
